Move ConvBNReLU to sg.modules.__init__ temporary.
